requirements.txt
celery==5.2.2
Django==3.2.12
#django-filter==2.4.0
django-settings-export==1.2.1
djangorestframework==3.13.1
djangorestframework-jsonp==1.0.2
djangorestframework-simplejwt==5.0.0
djangorestframework-xml==2.0.0
djangorestframework-yaml==2.0.0
PyJWT==2.3.0
pymongo==3.12.3
python-memcached==1.59
pytz==2021.3
PyYAML==6.0
gunicorn==20.1.0
